Cache one run of the analysis to speed up tests
This introduces a "workdir" fixture that runs the pipeline up to the
ready-for-variant-calling BAM file (mapped.calling.bam). The fixture
is scoped at module level, that is, it runs only once, allowing us to
re-use these already computed parts of the pipeline in some of the
tests.

Tests run in ~1 minute now instead of over 3 minutes.